Optimize pipeline creation
This commit attempts to optimize pipeline creation by reusing the
base txt2img pipeline to maintain four different pipelines ready
to use at all times, thus reducing the places in the code that
require rebuilding the generation pipeline.

The available generation pipelines are:
- StableDiffusionPipeline
- StableDiffusionImg2ImgPipeline
- StableDiffusionControlNetPipeline
- StableDiffusionControlNetImg2ImgPipeline

For SDXL the corresponding pipelines are created in a similar
fashion.

This commit also tries to ensure that the garbage collector is
called when rebuilding pipelines by removing all pipeline
references when rebuilding, including a pipeline reference in the
LoRA code that was preventing the garbage collector from releasing
the pipeline memory.